🐼
not necessary

+++ main.c
@@ -7,7 +7,7 @@
 #include <stdint.h>
 #include<sys/time.h>
 
-long long tseed(void) {
+long tseed(void) {
     struct timeval tv;
     gettimeofday(&tv,NULL);
     return (((long long)tv.tv_sec)*1000)+(tv.tv_usec/1000);
